Merged PR 6841: Allowing UI to post collection on save of first document
This is quite untested as the API does not allow this behavior yet. I spun up a debug version and made sure the call was coming through at the correct time, but I have not tested everything. This is just to get the code out there to be looked at

Related work items: #149513